Moved dataset_enums to datasets_base
