+++ backend/dataall/modules/s3_datasets/services/dataset_table_service.py
@@ -16,7 +16,7 @@ from dataall.modules.s3_datasets.services.dataset_permissions import (
     SYNC_DATASET,
 )
 from dataall.modules.s3_datasets.db.dataset_repositories import DatasetRepository
-from dataall.modules.s3_datasets.services.datasets_enums import ConfidentialityClassification
+from dataall.modules.datasets_base.services.datasets_enums import ConfidentialityClassification
 from dataall.modules.s3_datasets.db.dataset_models import DatasetTable, Dataset
 from dataall.modules.s3_datasets.services.dataset_permissions import (
     PREVIEW_DATASET_TABLE,
